+++ NetDoc/Program.cs
@@ -117,18 +117,4 @@ namespace NetDoc
                 .Distinct(new FileNameOnlyComparer());
         }
     }
-
-    internal class FileNameOnlyComparer : IEqualityComparer<string?>
-    {
-        public bool Equals(string? left, string? right)
-        {
-            if (left == null) return right == null;
-            return Path.GetFileName(left).Equals(Path.GetFileName(right), StringComparison.OrdinalIgnoreCase);
-        }
-
-        public int GetHashCode(string? obj)
-        {
-            return Path.GetFileName(obj)?.GetHashCode() ?? -1;
-        }
-    }
 }
